Adding v1 of crate template
The template is used for creating new repositories in
rust-vmm.

It contains the following files:
- README.md with a few tips for writing the crate readme
- CODEOWNERS
- coverage_config.json, .gitmodules and rust-vmm-ci which are used
  for running the same CI on all rust-vmm components
- LICENSE-APACHE
- a dummy src/lib.rs
- Cargo.toml

The BSD-Clause-3 license is not commited to this template on purpose
because the text of the license includes the Copyright. The Copyright
is different depending on the contents of the crates as well as the
authors.

Signed-off-by: Andreea Florescu <fandree@amazon.com>