Use proc macros to opt-in/opt-out serde
A common requirement for virtual machine snapshot, live migration and
live upgrading is to serialize and deserialize VMM internal objects and
data structures. The rust community has the excellent serde/serde_derive
framework for data serialization/deserialization. But the
serde/serde_derive framework have very heavy dependency chain, and it's
a common concern of the rust-vmm project to reduce dependency chain as
small as possible. So this helper crate uses features to opt-in data
serialization/deserialization functionality.

Signed-off-by: Liu Jiang <gerry@linux.alibaba.com>