Add LICENSE-BSD-3-Clause file
Some crates written from scratch, may want to use a BSD-3-Clause
license with a generic copyright. Let's add a BSD-3-Clause license
file in this template with a generic "The rust-vmm authors"
copyright that can be used as is or changed.

Update the readme about this new license file.

Signed-off-by: Stefano Garzarella <sgarzare@redhat.com>