Commit Cargo.lock by default
We audit crates for vulnerabilities on CI already. Lacking a lockfile
means the audit is meaningless if the dependency versions can change
between subsequent runs because of a e.g. minor or patch semver fix of a
security vulnerability.

There are advantages and disadvantages to having deterministic
dependencies:

Advantages:

- Aforementioned auditing
- Bisectability
- More predictable behavior when binaries are installed with `cargo
  install --locked ...`.

Disadvantages:

- Cargo's dependency resolution can end up duplicating transitive
  dependencies in downstream crates which use a `rust-vmm` dependency
  that shares dependencies with more than one of the others. This can be
  inspected with `cargo-tree`.
- Requires extra maintainer attention and effort to keep the lockfile
  up-to-date.

The Rust Cargo team currently recommends against committing the lockfile
for library crates in version control, but plans to reverse this
  recommendation in the future. You can read the new Cargo
  recommendations in the Nightly version docs:

* https://doc.rust-lang.org/nightly/cargo/faq.html#why-have-cargolock-in-version-control
* Web snapshot in Internet Archive: http://web.archive.org/web/20230921033901/https://doc.rust-lang.org/nightly/cargo/faq.html#why-have-cargolock-in-version-control

Signed-off-by: Manos Pitsidianakis <manos.pitsidianakis@linaro.org>